Refactor timer management and add profile sanity check
- Renamed timer variables for clarity from vanTimers to countTimers in shrimpmgr.cpp.
- Updated timer handling logic to ensure correct indexing and prevent overflow.
- Implemented a new function sanity_check_profile() in webserver.cpp to validate profile configurations and check for overlaps in cycles and horarios.
- Added error handling to prevent applying invalid profiles.
- Introduced new main.c and wifi.c files to set up the Mongoose web server and Wi-Fi connection, respectively.
- Configured CMakeLists.txt to include new source files for the project.

+++ main/inc/typedef.h
@@ -170,8 +170,8 @@ PV1V
 // ============================================================================
 
 typedef struct {
-    uint8_t hourStart;
-    uint8_t horarioLen;
+    float hourStart;
+    float horarioLen;
     uint8_t pwmDuty;
 } horario_t;
